requirements.txt
# Datenverarbeitung
pandas
numpy
scipy

# Visualisierung
matplotlib
seaborn

# Interaktive Charts (Dashboard)
plotly 
streamlit

# Machine Learning und Skalierung
scikit-learn
holidays
cmdstanpy
prophet
lightgbm
shap
wheel

# API & Hilfsmittel
spotipy
python-dotenv
requests
tqdm
google-generativeai
# Entwicklung
jupyter 
